refactor: utils dir structure
