+++ src/components/App.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types'
 import { ErrorBoundary } from 'react-error-boundary'
 import { toast } from 'react-toastify'
 import RenderForm from './RenderForm'
-import { readFromJSONFile } from '../utilities/fileUtils'
-import { preProcessSchema } from '../utilities/schemaHandlers'
-import { fetchSchemasfromS3, findLatestSchemas, parseAndFilterSchemas, findSchemaFromData } from '../utilities/schemaFetchers'
+import { readFromJSONFile } from '../utils/helpers/file.helpers'
+import { preProcessSchema } from '../utils/helpers/schema-handlers.helpers'
+import { fetchSchemasfromS3, findLatestSchemas, parseAndFilterSchemas, findSchemaFromData } from '../utils/helpers/schema-fetchers.helpers'
 import Toolbar from './Toolbar'
 import styles from './App.module.css'
 import { nanoid } from 'nanoid'
